Backend/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pycparser==2.23
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
drf-spectacular==0.29.0
google-genai

djangorestframework-simplejwt   #토큰 refresh를 위한 패키지
requests
#유비콘 추가 패키지
uvicorn[standard]
django
mysqlclient
python-dateutil
pycryptodome>=3.20.0
